Apply suggestions from code review

+++ Objects/rangeobject.c
@@ -1153,7 +1153,7 @@ range_iter(PyObject *seq)
 
     if (r->start == _PyLong_GetZero() && r->step == _PyLong_GetOne() ) {
         /* fast path for one argument range */
-        return fast_range_iter(0, lstop, 1, Py_MAX(0, (long)lstop));
+        return fast_range_iter(0, lstop, 1, Py_MAX(0, lstop));
     }
 
     lstart = PyLong_AsLong(r->start);
